feat: Hide docs voting on error or loading
