+++ apps/web/src/components/docs/Docs.tsx
@@ -1,14 +1,13 @@
-import { colors, IconInfoOutline, IconOutlineWarning, Tabs, Tooltip } from '@novu/design-system';
 import { Accordion, Alert, Code, Loader, Paper } from '@mantine/core';
-import { PropsWithChildren, ReactNode, useEffect, useMemo } from 'react';
-import * as mdxBundler from 'mdx-bundler/client';
-import { Highlight } from './Highlight';
-import { useQuery } from '@tanstack/react-query';
-import { useSegment } from '../providers/SegmentProvider';
-import { Center, Flex, Grid, GridItem, styled, VStack } from '@novu/novui/jsx';
+import { colors, IconInfoOutline, IconOutlineWarning, Tabs, Tooltip } from '@novu/design-system';
 import { css } from '@novu/novui/css';
+import { Center, Flex, Grid, GridItem, styled, VStack } from '@novu/novui/jsx';
 import { text, title as RTitle } from '@novu/novui/recipes';
-import { DOCS_URL, MDX_URL, MINTLIFY_IMAGE_URL } from './docs.const';
+import * as mdxBundler from 'mdx-bundler/client';
+import { PropsWithChildren, ReactNode, useEffect, useMemo } from 'react';
+import { DOCS_URL, MINTLIFY_IMAGE_URL } from './docs.const';
+import { Highlight } from './Highlight';
+import { DocsQueryResults } from './useLoadDocs';
 
 const Text = styled('p', text);
 const LiText = styled('span', text);
@@ -20,31 +19,18 @@ const getMDXComponent = mdxBundler.getMDXComponent;
 
 const DOCS_WRAPPER_ELEMENT_ID = 'embedded-docs';
 
+type DocsProps = PropsWithChildren<
+  DocsQueryResults & {
+    isLoading?: boolean;
+    actions: ReactNode;
+  }
+>;
+
 /*
- *Render the mdx for our mintlify docs inside of the web.
- *Fetching the compiled mdx from another service and then try to map the markdown to react components.
+ * Render the mdx for our mintlify docs inside of the web.
+ * Fetching the compiled mdx from another service and then try to map the markdown to react components.
  */
-export const Docs = ({ path = '', children, actions }: PropsWithChildren<{ path?: string; actions: ReactNode }>) => {
-  const segment = useSegment();
-
-  const { isLoading, data: { code, title, description } = { code: '', title: '', description: '' } } = useQuery<{
-    code: string;
-    title: string;
-    description: string;
-  }>(['docs', path], async () => {
-    const response = await fetch(MDX_URL + path);
-    const json = await response.json();
-
-    return json;
-  });
-
-  useEffect(() => {
-    segment.track('Inline docs opened', {
-      documentationPage: path,
-      pageURL: window.location.href,
-    });
-  }, [path, segment]);
-
+export const Docs = ({ code = '', description = '', title = '', isLoading, children, actions }: DocsProps) => {
   const Component = useMemo(() => {
     if (code.length === 0) {
       return null;
